+++ documentation/release_notes.md
@@ -3,7 +3,6 @@
 
 ## Changes in 3.3
 
-
 **New**
 - MELT user guide allows for downloading track data directly via the documentation page.
 - `TrainTestSplit` (with stratification) for any Java object and specialized for alignments (`TrainTestSplitAlignment`).
@@ -18,6 +17,9 @@
 - New mode for `TopXFilter`: `SOURCE_AND_TARGET`
 - Python server uses now gensim 4.x and flask 2.x ([issue #67](https://github.com/dwslab/melt/issues/67)).
 - Python server can be shut down via `/shutdown`.
+
+**New Tracks**
+- Process Matching (2017)
 
 ## Changes in 3.2